python_services/requirements.txt
# Unified Python Services for Migration Manager
# All dependencies for PDF, Email, and Document processing

# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0

# PDF Processing
PyPDF2>=3.0.0
pdf2image>=1.16.0
Pillow>=9.0.0
reportlab>=4.0.0

# Email Processing
extract-msg>=0.40.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
html5lib>=1.1

# Document Conversion
python-docx>=1.0.0
openpyxl>=3.1.0

# Utilities
python-multipart>=0.0.6
aiofiles>=23.0.0
python-dateutil>=2.8.0

# Security & Validation
bleach>=6.0.0
validators>=0.20.0

# Logging & Monitoring
loguru>=0.7.0

